polymarket-copy-trading/dashboard_requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0
asyncio-mqtt>=0.16.0
